Refactoring the code and adding the script on tmpl file to make it suitable for code
