+++ cmd/web/routes.go
@@ -3,8 +3,8 @@ package main
 import (
 	"net/http"
 
-	"github.com/Charik-Goyal/bookings/pkg/config"
-	"github.com/Charik-Goyal/bookings/pkg/handlers"
+	"github.com/Charik-Goyal/bookings/internal/pkg/config"
+	"github.com/Charik-Goyal/bookings/internal/pkg/handlers"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 )
@@ -21,8 +21,11 @@ func routes(app *config.AppConfig) http.Handler {
 	mux.Get("/about", handlers.Repo.About)
 	mux.Get("/generals-quarters", handlers.Repo.Generals)
 	mux.Get("/majors-suite", handlers.Repo.Majors)
+
 	mux.Get("/search-availability", handlers.Repo.Availability)
 	mux.Post("/search-availability", handlers.Repo.PostAvailability)
+	mux.Post("/search-availability-json", handlers.Repo.AvailabilityJSON)
+
 	mux.Get("/contact", handlers.Repo.Contact)
 
 	mux.Get("/make-reservation", handlers.Repo.Reservation)
