+++ internal/pkg/handlers/handlers.go
@@ -1,12 +1,14 @@
 package handlers
 
 import (
+	"encoding/json"
 	"fmt"
+	"log"
 	"net/http"
 
-	"github.com/Charik-Goyal/bookings/pkg/config"
-	"github.com/Charik-Goyal/bookings/pkg/models"
-	"github.com/Charik-Goyal/bookings/pkg/render"
+	"github.com/Charik-Goyal/bookings/internal/pkg/config"
+	"github.com/Charik-Goyal/bookings/internal/pkg/models"
+	"github.com/Charik-Goyal/bookings/internal/pkg/render"
 )
 
 // Repo the repository used by the handlers
@@ -77,6 +79,27 @@ func (m *Repository) PostAvailability(w http.ResponseWriter, r *http.Request) {
 	end := r.Form.Get("end")
 
 	w.Write([]byte(fmt.Sprintf("start date is %s, end date is %s", start, end)))
+}
+
+type jsonResponse struct {
+	OK      bool   `json:"ok"`
+	Message string `json:"message"`
+}
+
+// AvailabilityJson handles request for availability and send json request
+func (m *Repository) AvailabilityJSON(w http.ResponseWriter, r *http.Request) {
+	resp := jsonResponse{
+		OK:      false,
+		Message: "Available!",
+	}
+
+	out, err := json.MarshalIndent(resp, "", "    ")
+	if err != nil {
+		log.Println(err)
+	}
+
+	w.Header().Set("Content-Type", "application/json") //Header tells the web-browser what type of response i am sending
+	w.Write(out)
 }
 
 // Contact renders the contact page
