+++ templates/base.layout.tmpl
@@ -182,19 +182,16 @@
                         title = "",
                     } = c
 
-                    const { value: formValues } = await Swal.fire({
+                    const { value: result } = await Swal.fire({
                         title: title,
                         html: msg,
                         backdrop: false,
                         focusConfirm: true,
                         showCancelButton: true,
                         willOpen: () => {
-                            const elem = document.getElementById("reservation-dates-modal")
-                            const rp = new DateRangePicker(elem, {
-                                format: 'yyyy-mm-dd',
-                                showOnFocus: true,
-                                buttonClass: 'btn',
-                            })
+                            if (c.willOpen !== undefined) {
+                                c.willOpen()
+                            }
                         },
                         preConfirm: () => {
                             return [
@@ -203,13 +200,24 @@
                             ]
                         },
                         didOpen: () => {
-                            document.getElementById('start').removeAttribute('disabled')
-                            document.getElementById('end').removeAttribute('disabled')
+                            if (c.didOpen !== undefined) {
+                                c.didOpen()
+                            }
                         }
                     })
 
-                    if (formValues) {
-                        Swal.fire(JSON.stringify(formValues))
+                    if (result) {
+                        if (result.dismiss !== Swal.DismissReason.cancel) {
+                            if (result.value !== "") {
+                                if (c.callback !== undefined) {
+                                    c.callback(result);
+                                }
+                            } else {
+                                c.callback(false);
+                            }
+                        } else {
+                            c.callback(false);
+                        }
                     }
                 }
