+++ templates/generals.page.tmpl
@@ -50,7 +50,41 @@
                     </div>
                 </form>
             `
-        attention.custom({ msg: html, title: "Pick date" })
+        attention.custom({
+            msg: html,
+            title: "Pick date",
+
+            willOpen: () => {
+                const elem = document.getElementById("reservation-dates-modal")
+                const rp = new DateRangePicker(elem, {
+                    format: 'yyyy-mm-dd',
+                    showOnFocus: true,
+                    buttonClass: 'btn',
+                })
+            },
+
+            didOpen: () => {
+                document.getElementById('start').removeAttribute('disabled')
+                document.getElementById('end').removeAttribute('disabled')
+            },
+
+            callback: function (result) {
+                console.log("called")
+
+                let form = document.getElementById("check-availability-form")
+                let formData = new FormData(form)
+                formData.append("csrf_token", "{{.CSRFToken}}")
+
+                fetch('/search-availability-json', {
+                    method: "post",
+                    body: formData,
+                })
+                    .then(response => response.json())
+                    .then(data => {
+                        console.log(data)
+                    })
+            }
+        })
     })
 </script>
 {{end}}
